+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: New `setting` prop for `Text` component and new `lineHeight` options.
+
 ## [2.78.0] - 2020-07-06
 
 Feature: